Refactor game data action and presentation modules
