+++ src/game_data.c
@@ -223,31 +223,11 @@ slayer3d_input_manager *runtime_input(const slayer3d_game_data_runtime *runtime)
 
 void actor_set_position(slayer3d_registered_actor *actor, slayer3d_vec3 position);
 void copy_property_value(slayer3d_properties *target, const char *key, const slayer3d_value *value);
-actor_pool_runtime *find_actor_pool(slayer3d_game_data_runtime *runtime, const char *name);
-static const actor_pool_runtime *find_actor_pool_const(const slayer3d_game_data_runtime *runtime, const char *name);
-static actor_pool_runtime *find_actor_pool_for_actor(slayer3d_game_data_runtime *runtime, const char *actor_name,
-                                                     int *out_index);
-const actor_pool_runtime *find_actor_pool_for_actor_const(const slayer3d_game_data_runtime *runtime,
-                                                          const char *actor_name, int *out_index);
 bool actor_pool_in_scene(const actor_pool_runtime *pool, const char *scene_name);
-slayer3d_registered_actor *actor_pool_allocate(slayer3d_game_data_runtime *runtime, actor_pool_runtime *pool,
-                                               int *out_index);
-static actor_lifecycle_state actor_pool_lifecycle_state(const actor_pool_runtime *pool, int index);
-void actor_pool_set_lifecycle_state(actor_pool_runtime *pool, slayer3d_registered_actor *actor, int index,
-                                    actor_lifecycle_state state);
-bool actor_pool_actor_is_active(const actor_pool_runtime *pool, const slayer3d_registered_actor *actor, int index);
-static bool actor_pool_actor_is_available(const actor_pool_runtime *pool, const slayer3d_registered_actor *actor,
-                                          int index);
 static int actor_pool_active_count(const slayer3d_game_data_runtime *runtime, const actor_pool_runtime *pool);
 static int actor_pool_available_count(const slayer3d_game_data_runtime *runtime, const actor_pool_runtime *pool);
-void actor_pool_note_spawn_attempt(actor_pool_runtime *pool);
-void actor_pool_note_spawn_success(slayer3d_game_data_runtime *runtime, actor_pool_runtime *pool);
-void actor_pool_note_spawn_failure(actor_pool_runtime *pool, const char *reason);
-bool initialize_pooled_actor(actor_pool_runtime *pool, slayer3d_registered_actor *actor, int index, bool active);
 static bool actor_pool_initialize_slot(slayer3d_game_data_runtime *runtime, actor_pool_runtime *pool, int index,
                                        bool active);
-bool actor_pool_request_despawn(slayer3d_game_data_runtime *runtime, actor_pool_runtime *pool,
-                                slayer3d_registered_actor *actor, int index, const char *reason);
 static bool apply_actor_pool_scene_exit_policies(slayer3d_game_data_runtime *runtime, const char *from_scene,
                                                  const char *to_scene);
 bool actor_matches_target_filter(const slayer3d_game_data_runtime *runtime, const slayer3d_registered_actor *target,
@@ -262,15 +242,11 @@ void actor_lifecycle_defer_end(slayer3d_game_data_runtime *runtime);
 
 #include "game_data/game_data_render_runtime.inc"
 
-#include "game_data/game_data_ui_animation_runtime.inc"
-
 #include "game_data/game_data_scene_flow_runtime.inc"
 
 #include "game_data/game_data_menu_ui.inc"
 
 #include "game_data/game_data_actors_input.inc"
-
-#include "game_data/game_data_actions.inc"
 
 #include "game_data/game_data_scene_load_runtime.inc"
